Split cmd/safe.go into pending, display, batch, and tx-builder files
Move-only, same package cmd. Commands and deploy helpers stay in
safe.go; no Jarvis behavior change.
